JOLICLOUD: eeepc-laptop: Allow direct access via i2c for manipulating FSB
The asus-eee module was a third-party kernel module that was created when
the EeePC first came out. It allowed direct low-level access to the
temperature, fans, and the front-side bus. While the module itself was
never imported into the kernel directly, the temperature and fan read-outs
was made available, but the FSB access was missing..

Manipulating the FSB was a technique some netbook distributions did to
squeeze as much power as they could out of the EeePC.

The FSB is manipulated through the I2C-I801 interface.

Signed-off-by: Adam McDaniel <adam@jolicloud.org>