JOLICLOUD: broadcom: uuencode upstream shared objects
When building the kernel, the process that generates the source code
cannot create a diff from upstream if there are binaries added afterwards.
For now, uuencode the binaries so they can be represented in the diff.

This solution is hacky, and should be removed with something more elegant,
like 100% open-sourced drivers.