JOLICLOUD: poulsbo: Capped NULL pointer if "integrated sync" not supported
A bug in the poulsbo driver causes a kernel NULL pointer dereference if
integrated sync or stereo mode was not supported by the display.
This was introduced by drm_mode_detailed(), but the subroutines that called
function did not handle the condition gracefully.

End result is that some poulsbo users had displays that crashed on X
startup.

This issue was originally identified and fixed by Mandriva bug
https://qa.mandriva.com/show_bug.cgi?id=51562

Signed-off-by: Adam McDaniel <adam@jolicloud.org>