intel_rng: Attempt to register RNG, even if firmware auto-probe fails
Some hardware chipsets have an Intel RNG available, but intel_rng.ko fails
to initialize it because the firmware access procedure was invalid.
This hardware does not require a firmware init procedure, and will run just
fine without it.

On top of that, the failing firmware init process adds 0.5-1.0 seconds to
the boot time due to the use of stop_machine().  While unavoidable, we can
at least configure the driver to skip the firmware load attempt on devices
we know do not require it.

The Intel Atom N270 is one such chipset that will benifit from this.

Signed-off-by: Adam McDaniel <adam@jolicloud.org>