AppArmor: fix refcount order bug that can trigger during replacement
AppArmor uses lazy reference counting on chains of reference to reduce
refcounting overhead.  When loading a replacement profile races with
a task replacing the creds current reference it is possible that
cxt->reference is the reference at the head of the replacement chain,
dropping this reference can actually trigger the lose of the last
profile in the chain if there are no more references to it.

Signed-off-by: John Johansen <john.johansen@canonical.com>