AppArmor: fix typo in scrubbing environment variable warning
The debug warning of scrubbing the environment variables is used by
userspace to detect when exec and permission failures may be caused
by environment variable scrubbing like when firefox execs java
as in bug http://bugs.launchpad.net/bugs/484148.  Having it mispelled
makes it difficult to grep for in the logs.

Signed-off-by: John Johansen <john.johansen@canonical.com>