AppArmor: make the global side the correct type
The global sid type was not not properly updated when the sid was transitioned
from a u16:u16 pair of global and user sid to a single u32 sid.  This
causes the sid to wrap, this won't cause problems for mediation, but could
conceivabley cause problems for an extremely long lived learning session
where profile are replaced a lot.

Signed-off-by: John Johansen <john.johansen@canonical.com>