AppArmor: The LSM is expecting the string len rather than the allocation size
AppArmor getprocattr was returning the wrong size for name value for
unconfined tasks, as it return the size of memory allocated - 1 (\0 is omitted)
instead of the size of the string.  In the case of unconfined tasks the mode
string is not output so the return size needs to be adjusted appropriately.

Signed-off-by: Kees Cook <kees.cook@canonical.com>
Signed-off-by: John Johansen <john.johansen@canonical.com>