aufs: possible bugfix, sbinfo lock in deleting inode, core
(A commit in a series of introducing pid map/tree and making sure to
acquire sbinfo lock in deleting inode. While every commit is
git-bisect-able, you shoule read all commits in the series since a
single commit may have less meaning.)

s_umount rwsem in struct super_block prevents a race condition among
umount, remount and kswapd. It is good.
But if an inode is going to be deleted by other than kswapd, aufs may be
doing another operation which requires a lock for sbinfo. In this case,
si_noflush_read_trylock() in au_iinfo_fin() will not acquire the
lock. Before au_iinfo_fin() completes or during its operations, another
operation may release the lock and remount or a branch management
process may start. Here if the branch management process changes the
union members, then xino management in au_iinfo_fin() will not work
correctly.

In order to fix this potential problem, there another bad approch is
introduced which uses a bitmap to mark the pid which acquired the sbinfo
lock. In au_iinfo_fin(), if the bit is set, then the function will not
try acquiring the lock.
To support the pid larger than PID_MAX_DEFAULT, sbinfo prepares a radix
tree too.

With this commit, we can remove si_noflush_read_trylock(). But it will
be necessary in aufs2-31 branch, so leave it now.

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>