UBUNTU: [Upstream] Input: evdev - never leave the client buffer empty after write
When the client buffer is very small and wraps around a lot, it may
well be that a write increases the head such that head == tail. If
this happens between the point where a poll is triggered and the
actual data is being read, there will be no data to read. This is
confusing to applications, which might end up closing the file.

This patch solves the problem by making sure the client buffer is
never empty after writing to it.

Currently in upstream maintainer's input tree and queued for v2.6.36:
git://git.kernel.org/pub/scm/linux/kernel/git/dtor/input.git next
commit e725a4945d6eedd400dd5d0ead293d980a2f76ec

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>
Signed-off-by: Dmitry Torokhov <dtor@mail.ru>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>