AppArmor: Fix oops when auditing the addition of profile namespace
The name of the namespace is no longer dynamically allocated as part of
the profile unpacking, so it should not be freed.  Also the when a
new namespace is created the ref count is wrong as a ref count is needed
for the list, as well as the returned pointer.
Part of lp #408454

Signed-off-by: John Johansen <john.johansen@canonical.com>