AppArmor: change network auditing over to use same structs as lsm_audit
current networking code needs to be expaned, to provide more information
make it similar to lsm_audit network auditing so that it is easy to
convert.

Also drop revalidation in the post_create hook, it makes no sense to
have it here.  It will get reintroduced again later in the new
networking hooks.

Signed-off-by: John Johansen <john.johansen@canonical.com>