AppArmor: Fix mediation of created paths that look like "deleted" paths
Under some filesystems (that provide their own dentry instantiation)
the allocated dentry can be flag as unhashed (same as a deleted dentry)
until the d_inode is allocated.  Because of the location of the
security_path hooks AppArmor (and __d_path) sees the requested
path as deleted, and fails to allow the creation of the file/node,
as it is not covered by deleted file delegation.
Detect path dentries without allocated d_inodes and treat them
as regular file paths.

Signed-off-by: John Johansen <john.johansen@canonical.com>