AppArmor: Add debug output when asking for secure exec.
Secure exec sets flags so that glibc will clear environment variables
that can be dangerous.  However this can cause applications to fail
in unexpected ways and there is no indication that AppArmor is asking
for the environment variables to be scrubbed because the exec doesn't
actually fail.

This makes debugging difficult and results in bugs being filed that
against AA that need to be fixed in policy.

Add a debugging statement that can be toggled by turning on debug mode
to allow quick checking of this.  In the future a new auditing mode
should be added for this type of event.

Signed-off-by: John Johansen <john.johansen@canonical.com>