AppArmor: Fix error code for when change_hat is done and there are no hats
Normally when change_hat is done and no hat is found ENOENT is returned
but when the profile has no hats it should return ECHILD.  This allows
the changehat api to bail out early instead of continuing to probe.

Update to return ECHILD when the profile has no hats or subprofiles.
This is not quiet right in that it doesn't distiguish between hats
(subprofiles with the hat flag) and ordinary subprofiles, but it is better
than what we had, and right the majority of the time currently.

Signed-off-by: John Johansen <john.johansen@canonical.com>