AppArmor: fix opps when child profiles are replaced
BugLink: http://bugs.launchpad.net/bugs/475619

Profile removal results in child profiles having their replacedby field
assigned an ERR_PTR.  Which will cause a null pointer oops.  This will
affect any application that has a profile in a hat, or a learning
mode subprofile when the base profile is removed.

The use of ERR_PTR used to be the way AppArmor would distinguish between a
removal and unconfined, this became invalid when the namespace->unconfined
profile was introduced.

After replacement the child process will change its context so that
its profile is set ERR_PTR, the ERR_PTR then gets filtered to a NULL ptr
which then causes the oops.

Instead of doing a minimal patch, cleanup the code that lead to this
bug.  Note, the newest_profile portion which tests for profile being
set to IS_ERR is now dead code and fixed in another patch.

Signed-off-by: John Johansen <john.johansen@canonical.com>