AppArmor: Rework AppArmor file delegation
File delegation requires both an id for the file and a mapping of the
delegations for x transitions.  So we need to add another delegate
field to the file permission structure.

The permission flags AA_MAY_DELEGATE and AA_EXEC_DELEGATE are not
required if we make the 0 index of the delegate field indicate not
delegate permission.

Since these permissions are not yet been activated this change is just
a necessary step towards enabling them.

Signed-off-by: John Johansen <john.johansen@canonical.com>