AppArmor: The caps cache is caching based off of task which is incorrect
The caps cache which is used to reduce duplicate capability messages
is caching based off of task pointer which is wrong.  While the stored
value is never dereferenced, there is no way to account for pointer reuse
if the task goes away.

Move to using a ref counted profile, which keeps the reference from
disappearing while cached and provides the granularity needed for user
space learning.

Signed-off-by: John Johansen <john.johansen@canonical.com>