AppArmor: fix null profile reference bug
The locking and refcounting around auto removing a null-profile is wrong.
It is supposed to automatically remove the profile when not referenced
by an object (task_context, file_context, ...), but only by system lists
etc.

However this leave a reference to the profile on the profile list that is
not reference counted, and may be read from the list before it is
properly removed.  For now fix this by having a proper list reference
for null-profiles.  This means that they do not automatically clean up
at this time.

Signed-off-by: John Johansen <john.johansen@canonical.com>