AppArmor: Add the option for the dfa to use the recursive default entry
The apparmor dfa has traditionally used default as a plain alternative
state, without recursion.  This is slightly simplier and can be faster
but it does not allow for as good of compression as states can not
be shared as well.

Add the ability to load a dfa where the default table entry is used
used recursively allowing for better table compression.

Signed-off-by: John Johansen <john.johansen@canonical.com>