AppArmor: Add a module parameter that controls in depth dfa verification
Give the option of turning off in depth dfa verification, at load time.
This will speed up loads a little by not walking every dfa state which
shouldn't technically be needed if the userspace is correctly generating
policy.  The basic check are left in place to catch most potential
problems.

Leave full check as the default so that skipping it is opt in only.

Signed-off-by: John Johansen <john.johansen@canonical.com>