AppArmor: Fix an oops introduced by the filter profile cleanup
When cleaning up filter_profile the change in bprm_set_creds introduced
an oops while fixing that ns my not be set to the newest profile namespace.
This happend because previously the ns was not set to the newest profile
and it was changed to use the newest profile namespace.  However
aa_confining_profile will return NULL for unconfined processes resulting
in an oops when setting ns.

Signed-off-by: John Johansen <john.johansen@canonical.com>