AppArmor: Fix the forced auditing and quieting of link permissions
An AppArmor file rule can be force audited or forced quiet so it doesn't
generate auditing messages.  In the case of link permissions the auditing
information is carried on the second part of the link pair.

Currently the forced auditing and quieting of link permissions is
failing because the link mediation function is setting the audit masks
for forced audit and quiet based off of the first lookup pair in the
link.  This however is not correct as the permissions associated  with
the first pair will never contain the masks for link permissions,
as we need to match both parts of the pair before we can know what
the final permissions set is.

Signed-off-by: John Johansen <john.johansen@canonical.com>