AppArmor: Fix allocation of the ns hname
When a namespace is created it is relative to the namespace it was
created from.  The name passed in for creation is the name relative
to that namespace.

However the hname should be relative to the global root.  Since the
name passed in is relative to the current namespace this would be
current ns->hname//name.

Signed-off-by: John Johansen <john.johansen@canonical.com>