AppArmor: move aa_dfa_null_transition out match, as it really isn't a dfa fn
aa_dfa_null_transition is just a special wrapper function that exists
soley to accomidate how apparmor does pair matching with in the dfa.
Move it into apparmor.h as that is where the lib headers are currently
and it will be used by more than file matching in the future.

Signed-off-by: John Johansen <john.johansen@canonical.com>