AppArmor: Remove references to sys
The AppArmor code has references to sys profile through out, this comes
from an earlier development of profile stacking that allowed two profiles
sys and user.  This is ability is not present in the current code and
when introduced will be replaced by a more generic form of profile stacking
so remove references to sys, which don't add to the code readability
and will need to be reworked before generic profile stacking is introduced.

Signed-off-by: John Johansen <john.johansen@canonical.com>