AppArmor: overlap link_subset permission
The link_subset flag can only occur in the second element of a pair
seperated by a null transition.  As such it can overlap another permission
in the permission set, as long as it doesn't need to be audited
seperately (which it doesn't).

This has the advantage of freeing up a bit in the permissions mask and
allowing the link_subset accept states to overlap with other accept
states allowing the minimized dfa to be smaller if the overlap can
be exploited.

Signed-off-by: John Johansen <john.johansen@canonical.com>