AppArmor: don't inherit rename profile sid if there is also an original profile
When AppArmor does a renaming replace it can replace two profiles.  The
profile it is renaming but also the any profile with the same name as
the new profile.

In this case the new_profile should inherit the sid the profile with the
same name and not the profile it is replacing.

Signed-off-by: John Johansen <john.johansen@canonical.com>