AppArmor: Reorder the file permissions and add the AA_MAY_DELETE perm
Reorder the internal mapping of file permission so that those that
apply to a file are together.

Also introduce the AA_MAY_DELETE permission which is implied by userspace
AA_MAY_WRITE.  Traditionally AppArmor avoided adding this permission
because the ability to write to a file meant that an attacker could
effectively delete a file by overwritting its contents.

However deleting a file can also leave the file name open for creation
by another user, which in some cases may be undesirable.

Signed-off-by: John Johansen <john.johansen@canonical.com>