AppArmor: Remove unnecessary fields from the file audit entries
The file audit entries don't need to contain as much information as they
do.  While the permission mask audit, quiet, kill fields are used in
determining whether an audit message is output they are not part of
the audit message it self.

Rework so that only file data that will be actually part of the audit
message is in the audit fields.

Signed-off-by: John Johansen <john.johansen@canonical.com>