AppArmor: Remove the ability to set capability
AppArmor 2.3 added the ability to set capabilities, as an experimental
means of raising a task privilege when attached to a specific profile.
This ability came with a strong warning about using it, and it has
never been properly finished.

With the completion of change_profile, and set capabilities not fully
implemented.  Set capabilities can be a vulnerability if used in profiles,
(which should not have been done except in experimental profiles).
The vulnerability is that any unconfined user is allowed to change_profile
into any profile on the system.  This is fine as long as profiles strictly
reduce privilege but this is not the case with set capability.

Since what set capabilities was designed to achieve can be mostly done
with pam_cap.so and fscaps, remove set capabilities.  It won't be
reintroduced unless its additional abilities have a compelling use case
and only if it is complete.

Signed-off-by: John Johansen <john.johansen@canonical.com>