AppArmor: Enable renaming replacement of profiles.
Allow doing a renaming replace which replaces a profile with a new profile
of a different name.  This replaces set profile so that the tools can
replace learning profiles with profiles that have the correct name.

When a renaming replacement occurs it will replace both the profile using
the renaming name as well as any profile using the profile name.

Signed-off-by: John Johansen <john.johansen@canonical.com>