AppArmor: use the cred referenced by the file's security field
The cred referenced by the file's security field has a reference to
the profile at the time the file is created instead of adding a second
refcount to the profile use the profile referenced by the file cred.

This also fixes a bug with the current implementation, where
security_dentry_open was being used to assign the profile label to the
file but in some cases this will result in an unlabelled file.  The
solution would be to do the file labelling in aa_alloc_file_context
but that is no longer needed now that we just use the cred.

Signed-off-by: John Johansen <john.johansen@canonical.com>