AppArmor: change how on exec is computed to remove potential ambiguity
Currently onexec does a series of two check to determine change on exec
permission at the onexec call.

change_profile onexec :namespace:profile
change_profile profile

And then repeats the test, to walk the dfa to find permissions for
the exec portion of the onexec permission pair.  However this does more
work than is needed and has case where it can fail.

if two different change_profile onexec rules are defined, one for
namespaces, and another for profiles, and each specifies a different
exec match.
  eg.
     change_profile /bin/bin1 -> ns1://profile1,
     change_profile /bin/bin2 -> profile2,

the first part of the pair will match the ns rule but not to the
profile2 rule.  This will mean the /bin/bin2 transition will
never get match.

Since onexec isn't currently used break the old encoding to fix this.
Change to a two rule requirement.  For a userspace rule of
  change_profile /bin/bin1 -> profile,

the kernel will get 2 rules mapped.  The first is a non-paired,
rule.  Used by change_profile to check permission, the second is a paired
rule (like before) but with the change_profile permission as the second
pair.  This avoids the previous failure as the executable will always
be matched against and then all change_profile permission linked to the
pair.

Signed-off-by: John Johansen <john.johansen@canonical.com>