aufs: bugfix, separate the workqueue for preprocessing mmap
variation of common AB-BA deadlock problem.

ProcessA:
- aufs_mmap
  + pre-process with using a workqueue
  + wait until return from the workqueue

Workqueue task for ProcessA:
- acquire aufs rwsem

Processb
- lookup or readdir in aufs
  + acquire aufs rwsem
  + assign a new inode number
  + write the value to the XINO file using a workqueue
  + wait until return from the workqueue

Since the workqueue handles the request one by one, both of processA and
B waits forever.

This bug was introduced by the commit
d986fa5 2010-03-08
	aufs: bugfix, another approach to keep the lock order of mmap_sem
which is the last added workqueue task.
And this is the only one task which acquires such lock in workqueue.
To fix it, introduce another workqueue which is for preprocessing mmap only.
This commit will make the approach more ugly, I don't have another option.

Reported-by: Oliver Welter <mail@oliwel.de>
Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>