AppArmor: Change audit_header from type= to apparmor=
The audit_header currently outputs itself as type= as it was designed
to be used inconjunction with syslog when auditd isn't being used.
With the move to lsm_audit the apparmor message type was replaced by
the generic AUDIT_AVC (1400), so the audit header has been turned on by
default.  However system audit type and the AppArmor audit header both
use the same name so each message has two different type= values.  Fix
this by renaming the apparmor type to apparmor=, and with this drop
APPARMOR_ from the type string as apparmor="APPARMOR_AUDIT" is redundant.

Signed-off-by: John Johansen <john.johansen@canonical.com>