AppArmor: Merge kvmalloc from apparmorfs and table alloc from match.c
kvmalloc and allocations for dfa tables are almost identical and it
makes no sense carrying two slight different versions.  Merge them
into a single lib routine.

This does break the general desire to make the dfa free of the rest
of AppArmor code but it is easy to break again as kvmalloc is a
generic memory allocation routine.

Signed-off-by: John Johansen <john.johansen@canonical.com>