aufs: bugfix, buffer overrun in au_whtmp_lkup()
Removing or renaming a dir in aufs causes an internal renaming to a
temporary unique name. To gurantee its uniquness, aufs uses a static
internal counter and put it into the unique name. Historically it is
printed in hexdecimal, but I had changed it decimal. By this change, the
generating unique name will be longer, and when the counter exceeds 4
digtis AND the target dir name has AUFS_MAX_NAMELEN characterns, then
the buffer overrun will happen.

Reported-by: Thayumanavar Sachithanantham <thayumj@gmail.com>
Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>