aufs: more verbose at deleting a branch
In deleting a branch, aufs searches the files in-use on that branch. If
any of them are found, aufs rejects deleting immediately due to the
branch is busy. It is basically equivalent to unmounting.
But users may want to know the path of the opened file.
This commit stops aufs to abondon the deleting process in verbose mode
when it finds that there are file(s) opened on that branch. After
printing the paths, aufs returns the error as usual. While the result is
unchanged, aufs has to do for some unnecessary works just to show the
paths.

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>