aufs: simplify and remove au_test_wkq()
In whplink() and xino_fwrite(), au_test_wkq() is called to make sure
that we have super user privilege. Essentially whplink() needs the
privilege to lookup/link the hidden pseudo-link dir, and xino_fwrite()
needs to break RLIMIT_FSIZE (in order to suppress SIGXFSZ).

This commit splits (make the granularity lower) the tests.
Also in the debug build, make sure the workqueue has the expected fsid
and RLIMIT_FSIZE.

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>