UBUNTU: SAUCE: (drop after 2.6.35) ARM: Using gpmc function to init nand flash.
The General Purpose Memory Controller (gpmc) initialization was
decoupled from flash init but the changes were not reflected
in omap3beagle_flash_init.

This is the board specific portion of patch:
  commit f450d86790ebf72ac93c7ea5addd6fa278aae64c
  Author: Sukumar Ghorai <s-ghorai@ti.com>
  Date:   Fri Jul 9 09:14:46 2010 +0000

    omap3 nand: fix issue in board file to detect nand

This patch is supposed to be part of 2.6.36 but only a subset is
used to avoid dealing with the substantial amount of gpmc changes
that will go in.  These changes should be brought in when we
rebase to 2.6.36.

BugLink: https://bugs.launchpad.net/bugs/608266

Signed-off-by: Mathieu Poirier <mathieu.poirier@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>